+Solucionados varios bugs, por ahora coloque que el stock parcial se cargue lo mismo que tiene el stock normal. Tenemos que hablar como lo vamos a tratar, si vamos a poner estados de ventas o que. +Por ahora cada vez que se hace una venta, se resta la cantidad del stock parcial de los productos. +Modificados varios campos en los modelos.

+++ MinimercadoAlfredo/Models/SaleLine.cs
@@ -17,9 +17,9 @@ namespace MinimercadoAlfredo.Models
 
         public decimal LinePrice { get; set; }
 
-        public int LineDiscount { get; set; }
+        public decimal LineDiscount { get; set; }
 
-        public int LineQuantity { get; set; }
+        public decimal LineQuantity { get; set; }
 
         public decimal LineTotal { get; set; }
